+++ src/main/java/com/ERPsystem/miningcompany/service/scm/ProcurementService.java
@@ -2,6 +2,7 @@ package com.ERPsystem.miningcompany.service.scm;
 
 import com.ERPsystem.miningcompany.Entity.scm.Procurement;
 import com.ERPsystem.miningcompany.Repository.scm.ProcurementRepository;
+import com.ERPsystem.miningcompany.controller.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,20 +22,26 @@ public class ProcurementService {
         return procurementRepository.findAll();
     }
 
-    public Optional<Procurement> getProcurementById(Long id) {
-        return procurementRepository.findById(id);
+    public Procurement getProcurementById(Long id) {
+        return procurementRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Procurement not found with id " + id));
     }
 
     public Procurement updateProcurement(Long id, Procurement procurement) {
         if (procurementRepository.existsById(id)) {
             procurement.setId(id);
             return procurementRepository.save(procurement);
+        } else {
+            throw new ResourceNotFoundException("Procurement not found with id " + id);
         }
-        return null;
     }
 
     public void deleteProcurement(Long id) {
-        procurementRepository.deleteById(id);
+        if (procurementRepository.existsById(id)) {
+            procurementRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Procurement not found with id " + id);
+        }
     }
 
 }
