+++ src/main/java/com/ERPsystem/miningcompany/service/scm/SupplierService.java
@@ -2,6 +2,7 @@ package com.ERPsystem.miningcompany.service.scm;
 
 import com.ERPsystem.miningcompany.Entity.scm.Supplier;
 import com.ERPsystem.miningcompany.Repository.scm.SupplierRepository;
+import com.ERPsystem.miningcompany.controller.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,20 +22,26 @@ public class SupplierService {
         return supplierRepository.findAll();
     }
 
-    public Optional<Supplier> getSupplierById(Long id) {
-        return supplierRepository.findById(id);
+    public Supplier getSupplierById(Long id) {
+        return supplierRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Supplier not found with id " + id));
     }
 
     public Supplier updateSupplier(Long id, Supplier supplier) {
         if (supplierRepository.existsById(id)) {
             supplier.setId(id);
             return supplierRepository.save(supplier);
+        } else {
+            throw new ResourceNotFoundException("Supplier not found with id " + id);
         }
-        return null;
     }
 
     public void deleteSupplier(Long id) {
-        supplierRepository.deleteById(id);
+        if (supplierRepository.existsById(id)) {
+            supplierRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Supplier not found with id " + id);
+        }
     }
 
 }
